refactor and fix glue/stream
