+++ Sources/Wrp/Core/Glue.swift
@@ -4,6 +4,7 @@ import WebKit
 
 public class WrpGlue {
     public var webView: WKWebView?
+    private var initialized: Bool = false
     private var queue: DeferStream<Data> = .init()
     private var sharedSequence: SharedAsyncSequence<AsyncStream<Data>>
     private let configuration: Configuration
@@ -27,11 +28,16 @@ public class WrpGlue {
         }
     }
 
-    public func close() {
+    public func tryReconnect(afterReconnect: (() -> Void)?) {
+        guard self.initialized == true else {
+            self.initialized = true
+            return
+        }
         self.queue.continuation.finish()
         self.configuration.logger.debug("Closed")
         self.queue = .init()
         self.sharedSequence = self.queue.stream.shared()
+        afterReconnect?()
     }
 
     public func registerWebView(_ webView: WKWebView) {
